refactor(hover-card): extract generic HoverAnchor from HoverCardAnchor
